DLS-228 Testen und Problemlösung in Zusammenhang von CMI AIS

+++ CMI/Contract/Messaging/IArchiveRecordRemoved.cs
@@ -2,7 +2,7 @@ namespace CMI.Contract.Messaging
 {
     public interface IArchiveRecordRemoved
     {
-        int MutationId { get; set; }
+        long MutationId { get; set; }
         bool ActionSuccessful { get; set; }
         string ErrorMessage { get; set; }
         string StackTrace { get; set; }
